backend/requirements.txt
fastapi
uvicorn[standard]
pytest
groq
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
